LayoutTests/imported/w3c/web-platform-tests/css/css-mixins/function-container-self-expected.txt


FAIL Custom Functions: @container can not query self assert_equals: expected "B" but got "--f()"
